Add basic tools functionality:
Previously the tools where added as a route/ui-view. Changed
this to be a directive instead. The directive is placed on
the map and when the button is clicked, an overlay with HTML
content is placed below.

I also created a stub back-end request to return a list of
tool objects as a JSON array. I expect to use this data to
generate the buttons for the actual tools within the popover.