Fix conflicts with directives:
The 'Print' and 'Save Map As' directives are clashing. the
scopes are being mixed up as are handler functions. Refactored
so there are no name conflicts. I had hoped to use vm for all
controller interaction but that doesn't work, the same with
scopes. All scopes and controllerAs definitions have been
re-named.